refactor: make new message format separate
