+++ packages/dds/tree/src/shared-tree-core/messageFormatV1ToV4.ts
@@ -8,9 +8,7 @@ import { type TSchema, Type } from "@sinclair/typebox";
 
 import { type EncodedRevisionTag, RevisionTagSchema, SessionIdSchema } from "../core/index.js";
 import type { JsonCompatibleReadOnly } from "../util/index.js";
-import type { EncodedBranchId } from "./branch.js";
 
-// XXX: Create a new message format instead of changing the existing one.
 /**
  * The format of messages that SharedTree sends and receives.
  */
@@ -18,7 +16,7 @@ export interface Message {
 	/**
 	 * The revision tag for the change in this message
 	 */
-	readonly revision?: EncodedRevisionTag;
+	readonly revision: EncodedRevisionTag;
 	/**
 	 * The stable ID that identifies the originator of the message.
 	 */
@@ -26,9 +24,7 @@ export interface Message {
 	/**
 	 * The changeset to be applied.
 	 */
-	readonly changeset?: JsonCompatibleReadOnly;
-
-	readonly branchId?: EncodedBranchId;
+	readonly changeset: JsonCompatibleReadOnly;
 
 	/**
 	 * The version of the message. This controls how the message is encoded.
@@ -43,9 +39,8 @@ export interface Message {
 // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
 export const Message = <ChangeSchema extends TSchema>(tChange: ChangeSchema) =>
 	Type.Object({
-		revision: Type.Optional(RevisionTagSchema),
+		revision: RevisionTagSchema,
 		originatorId: SessionIdSchema,
-		changeset: Type.Optional(tChange),
-		branchId: Type.Optional(Type.Number()),
+		changeset: tChange,
 		version: Type.Optional(Type.Number()),
 	});
