+++ packages/dds/tree/src/test/shared-tree-core/messageCodec.spec.ts
@@ -16,7 +16,7 @@ import { typeboxValidator } from "../../external-utilities/index.js";
 // eslint-disable-next-line import/no-internal-modules
 import { makeMessageCodec, makeMessageCodecs } from "../../shared-tree-core/messageCodecs.js";
 // eslint-disable-next-line import/no-internal-modules
-import type { Message } from "../../shared-tree-core/messageFormat.js";
+import type { Message } from "../../shared-tree-core/messageFormatV1ToV4.js";
 // eslint-disable-next-line import/no-internal-modules
 import type { DecodedMessage } from "../../shared-tree-core/messageTypes.js";
 import { TestChange } from "../testChange.js";
